requirements.txt
PyQt5==5.15.9
pymongo==4.6.0
python-dotenv==1.0.0